remoção de código duplicado.
