+++ chatbots/telegram.py
@@ -5,7 +5,6 @@ import urllib
 import requests
 
 from chatbots.base import BaseBot
-from chatbots.botenum import BotFlowEnum
 from chatbots.model.bot_model import BotDbConnection
 from .utils import edu_logger
 
@@ -54,16 +53,6 @@ class TelegramBot(BaseBot):
     #
     # Private
     #
-
-    def _check_flow(self):
-        """if text is equal to initial statuses, them back to begin."""
-        if self.text in [BotFlowEnum.QUAL_CARDAPIO.value,
-                         BotFlowEnum.AVALIAR_REFEICAO.value,
-                         BotFlowEnum.RECEBER_NOTIFICACAO.value]:
-            self._reset_flow(self.text)
-
-    def _reset_flow(self, text):
-        self.set_flow(flow_name=text, flow_step=BotFlowEnum.STEP_INITIAL.value)
 
     def _concat_buttons(self, keyboard_opts, url, show_once=True):
         # https://core.telegram.org/bots/api/#keyboardbutton
